board/pine64/rockpro64/linux.fragment: integrate stmmac driver
Currently, the stmmac driver is configured as a module in the defconfig, which
means that the network functionality is not available until the driver module
is manually inserted.

Use extra config fragment to integrate it directly into the kernel, ensuring that network
functionality is available immediately upon boot.

Signed-off-by: Aleksandr Makarov <aleksandr.o.makarov@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 54c82a237581dbed74aab2086381ae69b4626cc6)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>