fs/erofs: add LZMA compression, allow configurable compression level
Instead of having a bool option for enabling the LZ4HC compression, change the
Kconfig definition to a choice, which also adds more flexibility in the future
(e.g. for enabling ZSTD compression). The name of the old option was preserved
so defconfigs using it won't be affected.

Two more options for compression were added, the standard LZ4 (i.e. non-HC)
compression algorithm and LZMA.

Also for LZMA and LZ4HC, it's now possible to define compression level.
Previously it defaulted to 9, which generates images larger than squashfs with
comparable settings (which uses level 12 by default). For LZ4HC the default if
unset would be still 9.

Signed-off-by: Jan Čermák <sairon@sairon.cz>